# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Food Subscription & Ledger Management System API",
    "description": "Production-grade, Clean Architecture REST API for Food Subscription & Ledger Management System.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/subscriptions": {
      "post": {
        "tags": [
          "Subscription Engine"
        ],
        "summary": "Create customer subscription",
        "description": "Creates a new subscription (Monthly, Weekly, Trial, Custom).\nAutomatically calculates start/end dates, Sunday holiday skips, custom rate overrides, and initial invoice.",
        "operationId": "create_subscription_api_v1_subscriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SubscriptionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/subscriptions/{subscription_id}": {
      "get": {
        "tags": [
          "Subscription Engine"
        ],
        "summary": "Get subscription details",
        "description": "Retrieves full details of a subscription including pending extension counters and calculated end date.",
        "operationId": "get_subscription_api_v1_subscriptions__subscription_id__get",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Subscription ID",
              "title": "Subscription Id"
            },
            "description": "Subscription ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SubscriptionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_SubscriptionResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates whether the request was processed successfully",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubscriptionResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable response message",
            "default": ""
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "List of error messages or validation failures"
          }
        },
        "type": "object",
        "title": "ApiResponse[SubscriptionResponse]"
      },
...
      "FoodPreference": {
        "type": "string",
        "enum": [
          "veg",
          "non_veg"
        ],
        "title": "FoodPreference"
      },
      "FoodPreferenceSelection": {
        "properties": {
          "breakfast": {
            "$ref": "#/components/schemas/FoodPreference",
            "default": "veg"
          },
          "lunch": {
            "$ref": "#/components/schemas/FoodPreference",
            "default": "non_veg"
          },
          "dinner": {
            "$ref": "#/components/schemas/FoodPreference",
            "default": "veg"
          }
        },
        "type": "object",
        "title": "FoodPreferenceSelection"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MealSelection": {
        "properties": {
          "breakfast": {
            "type": "boolean",
            "title": "Breakfast",
            "description": "Enable breakfast delivery",
            "default": true
          },
          "lunch": {
            "type": "boolean",
            "title": "Lunch",
            "description": "Enable lunch delivery",
            "default": true
          },
          "dinner": {
            "type": "boolean",
            "title": "Dinner",
            "description": "Enable dinner delivery",
            "default": true
          }
        },
        "type": "object",
        "title": "MealSelection"
      },
...
      "SubscriptionCreate": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "subscription_type": {
            "$ref": "#/components/schemas/SubscriptionType"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "meals": {
            "$ref": "#/components/schemas/MealSelection"
          },
          "preferences": {
            "$ref": "#/components/schemas/FoodPreferenceSelection"
          },
          "rates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubscriptionRates"
              },
              {
                "type": "null"
              }
            ]
          },
          "custom_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Days",
            "description": "Number of days if type is CUSTOM"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "subscription_type",
          "start_date",
          "meals",
          "preferences"
        ],
        "title": "SubscriptionCreate"
      },
      "SubscriptionRates": {
        "properties": {
          "breakfast_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breakfast Price",
            "description": "Override global breakfast price"
          },
          "lunch_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lunch Price",
            "description": "Override global lunch price"
          },
          "dinner_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dinner Price",
            "description": "Override global dinner price"
          },
          "delivery_charge": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Charge",
            "description": "Override global delivery charge per day"
          }
        },
        "type": "object",
        "title": "SubscriptionRates"
      },
      "SubscriptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "subscription_type": {
            "$ref": "#/components/schemas/SubscriptionType"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "original_end_date": {
            "type": "string",
            "format": "date",
            "title": "Original End Date"
          },
          "meals": {
            "$ref": "#/components/schemas/MealSelection"
          },
          "preferences": {
            "$ref": "#/components/schemas/FoodPreferenceSelection"
          },
          "rates": {
            "$ref": "#/components/schemas/SubscriptionRates"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "pending_breakfast_extensions": {
            "type": "integer",
            "title": "Pending Breakfast Extensions",
            "default": 0
          },
          "pending_lunch_extensions": {
            "type": "integer",
            "title": "Pending Lunch Extensions",
            "default": 0
          },
          "pending_dinner_extensions": {
            "type": "integer",
            "title": "Pending Dinner Extensions",
            "default": 0
          },
          "total_extended_days": {
            "type": "integer",
            "title": "Total Extended Days",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "subscription_type",
          "start_date",
          "end_date",
          "original_end_date",
          "meals",
          "preferences",
          "rates",
          "created_at",
          "updated_at"
        ],
        "title": "SubscriptionResponse"
      },
      "SubscriptionType": {
        "type": "string",
        "enum": [
          "monthly",
          "weekly",
          "trial",
          "custom"
        ],
        "title": "SubscriptionType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}